move rigid tests to separate file
